feat(Cooked Recipes): Users now can only access their own Cooked Recipes

+++ src/main/java/dev/leonardovcl/sweetcontrol/controllers/CookedRecipeController.java
@@ -11,6 +11,7 @@ import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -19,10 +20,13 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
 import dev.leonardovcl.sweetcontrol.model.CookedRecipe;
+import dev.leonardovcl.sweetcontrol.model.Recipe;
+import dev.leonardovcl.sweetcontrol.model.SecurityUser;
 import dev.leonardovcl.sweetcontrol.model.repository.CookedRecipeRepository;
 import dev.leonardovcl.sweetcontrol.model.repository.IngredientRepository;
 import dev.leonardovcl.sweetcontrol.model.repository.RecipeRepository;
 import dev.leonardovcl.sweetcontrol.model.repository.UsedInventoryRepository;
+import dev.leonardovcl.sweetcontrol.model.repository.UserRepository;
 import dev.leonardovcl.sweetcontrol.services.CookedRecipeService;
 
 @Controller
@@ -42,20 +46,24 @@ public class CookedRecipeController {
 	private RecipeRepository recipeRepository;
 	
 	@Autowired
+	private UserRepository userRepository;	
+	
+	@Autowired
 	private CookedRecipeService cookedRecipeService;
 	
 	@GetMapping
 	public String showCookedRecipes(
 			@RequestParam(value = "page", required = false,  defaultValue = "0") int page,
 			@RequestParam(value = "size", required = false, defaultValue = "5") int size,
-			@RequestParam(value = "idFilter", required = false) Long idFilter,
 			@RequestParam(value = "nameLike", required = false, defaultValue = "") String nameLike,
 			@RequestParam(value = "idIngredientFilter", required = false) Long idIngredientFilter,
+			@AuthenticationPrincipal SecurityUser securityUser,
 			Model model) {
+		
+		Long userId = userRepository.findByUsername(securityUser.getUsername()).get().getId();
 		
 		model.addAttribute("ingredientList", ingredientRepository.findAll());
 		
-		model.addAttribute("idFilter", idFilter);
 		model.addAttribute("nameLike", nameLike);
 		model.addAttribute("idIngredientFilter", idIngredientFilter);
 		
@@ -68,17 +76,19 @@ public class CookedRecipeController {
 		
 		List<CookedRecipe> cookedRecipeArrayList = new ArrayList<>();
 		
-		if(idFilter != null) {
+		if (!nameLike.isBlank() && idIngredientFilter == null) {
 			
-			cookedRecipeList = cookedRecipeRepository.findByRecipeEntryIdOrderByIdDesc(idFilter, pageable);
+			cookedRecipeArrayList = cookedRecipeService.findCookedRecipeByNameContaining(nameLike, userId);
 			
-		} else if (!nameLike.isBlank() && idIngredientFilter == null) {
+			cookedRecipeListHolder.setSource(cookedRecipeArrayList);
+			cookedRecipeListHolder.resort();
+			cookedRecipeListHolder.setPage(page);
 			
-			cookedRecipeList = cookedRecipeRepository.findByRecipeEntryNameContainingOrderByIdDesc(nameLike, pageable);
+			cookedRecipeList = new PageImpl<>(cookedRecipeListHolder.getPageList(), pageable, cookedRecipeArrayList.size());
 			
 		} else if (nameLike.isBlank() && idIngredientFilter != null) {
 			
-			cookedRecipeArrayList = cookedRecipeService.findCookedRecipeByIngredient(idIngredientFilter);
+			cookedRecipeArrayList = cookedRecipeService.findCookedRecipeByIngredient(idIngredientFilter, userId);
 			
 			cookedRecipeListHolder.setSource(cookedRecipeArrayList);
 			cookedRecipeListHolder.resort();
@@ -88,7 +98,7 @@ public class CookedRecipeController {
 		
 		} else if (!nameLike.isBlank() && idIngredientFilter != null) {
 			
-			cookedRecipeArrayList = cookedRecipeService.findCookedRecipeByIngredientAndNameContaining(idIngredientFilter, nameLike);
+			cookedRecipeArrayList = cookedRecipeService.findCookedRecipeByIngredientAndNameContaining(idIngredientFilter, nameLike, userId);
 			
 			cookedRecipeListHolder.setSource(cookedRecipeArrayList);
 			cookedRecipeListHolder.resort();
@@ -98,7 +108,13 @@ public class CookedRecipeController {
 			
 		} else {
 			
-			cookedRecipeList = cookedRecipeRepository.findAllByOrderByIdDesc(pageable);
+			cookedRecipeArrayList = cookedRecipeService.findAll(userId);
+			
+			cookedRecipeListHolder.setSource(cookedRecipeArrayList);
+			cookedRecipeListHolder.resort();
+			cookedRecipeListHolder.setPage(page);
+			
+			cookedRecipeList = new PageImpl<>(cookedRecipeListHolder.getPageList(), pageable, cookedRecipeArrayList.size());
 		}
 		
 		model.addAttribute("cookedRecipeList", cookedRecipeList);
@@ -107,20 +123,38 @@ public class CookedRecipeController {
 	}
 	
 	@GetMapping("/detail/{cookedRecipeId}")
-	public String detailCookedRecipe(@PathVariable("cookedRecipeId") Long cookedRecipeId, Model model) {
+	public String detailCookedRecipe(@PathVariable("cookedRecipeId") Long cookedRecipeId,
+									@AuthenticationPrincipal SecurityUser securityUser,
+									Model model) {
+		
+		Long userId = userRepository.findByUsername(securityUser.getUsername()).get().getId();
 		
 		CookedRecipe cookedRecipe = cookedRecipeRepository.findById(cookedRecipeId).get();
-		Long recipeId = cookedRecipe.getRecipeEntry().getId();
+		Recipe recipe = recipeRepository.findById(cookedRecipe.getRecipeEntry().getId()).get();
+		
+		if(recipe.getRecipeOwner().getId() != userId) {
+			return "error/accessDenied";
+		}
 		
 		model.addAttribute("cookedRecipe", cookedRecipe);
-		model.addAttribute("recipe", recipeRepository.findById(recipeId).get());
+		model.addAttribute("recipe", recipe);
 		model.addAttribute("usedInventoriesList", usedInventoryRepository.findByCookedRecipeEntryId(cookedRecipeId));
 		
 		return "cookedrecipes/cookedRecipesDetail";
 	}
 	
 	@GetMapping("/revert/{cookedRecipeId}")
-	public String revertCookedRecipe(@PathVariable("cookedRecipeId") Long cookedRecipeId) {
+	public String revertCookedRecipe(@PathVariable("cookedRecipeId") Long cookedRecipeId,
+									@AuthenticationPrincipal SecurityUser securityUser) {
+		
+		Long userId = userRepository.findByUsername(securityUser.getUsername()).get().getId();
+		
+		CookedRecipe cookedRecipe = cookedRecipeRepository.findById(cookedRecipeId).get();
+		Recipe recipe = recipeRepository.findById(cookedRecipe.getRecipeEntry().getId()).get();
+		
+		if(recipe.getRecipeOwner().getId() != userId) {
+			return "error/accessDenied";
+		}
 		
 		cookedRecipeService.revertCookedRecipe(cookedRecipeId);
 		
@@ -128,8 +162,18 @@ public class CookedRecipeController {
 	}
 	
 	@GetMapping("/delete/{cookedRecipeId}")
-	public String deleteCookedRecipe(@PathVariable("cookedRecipeId") Long cookedRecipeId) {
+	public String deleteCookedRecipe(@PathVariable("cookedRecipeId") Long cookedRecipeId,
+									@AuthenticationPrincipal SecurityUser securityUser) {
 	
+		Long userId = userRepository.findByUsername(securityUser.getUsername()).get().getId();
+		
+		CookedRecipe cookedRecipe = cookedRecipeRepository.findById(cookedRecipeId).get();
+		Recipe recipe = recipeRepository.findById(cookedRecipe.getRecipeEntry().getId()).get();
+		
+		if(recipe.getRecipeOwner().getId() != userId) {
+			return "error/accessDenied";
+		}
+		
 		cookedRecipeService.deleteCookedRecipe(cookedRecipeId);
 		
 		return "redirect:/cookedrecipes";
